# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ubhi Backend API
  version: 1.0.0
  description: >
    API for the Ubhi brand site. Powers the public storefront (workshops, shop,
    gallery, journal, bookings, orders, newsletter and snail-mail signups) and
    the back-office admin dashboard (auth, users, customers, settings).
    Stateless JWT auth; role + per-section permission model. Storage is
    pluggable (zero-config JSON file store by default, PostgreSQL when
    DATABASE_URL is set).
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8090
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        price:
          type: number
        qty:
          type: integer
    CreateOrderRequest:
      type: object
      required: [customer_email]
      properties:
        customer_email:
          type: string
          format: email
        customer_name:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        ship_street:
          type: string
        ship_city:
          type: string
        ship_postcode:
          type: string
        ship_country:
          type: string
        stripe_session_id:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_ref:
          type: string
        customer_name:
          type: string
        customer_email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        status:
          type: string
        ship_street:
          type: string
        ship_city:
          type: string
        ship_postcode:
          type: string
        ship_country:
          type: string
        stripe_session_id:
          type: string
          nullable: true
        paid:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Workshop:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        blurb:
          type: string
        session_date:
          type: string
        price:
          type: number
        spots:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        blurb:
          type: string
        price:
          type: number
        stock:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    GalleryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image:
          type: string
        caption:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    GalleryItemInput:
      type: object
      required: [image]
      properties:
        image:
          type: string
        caption:
          type: string
        published:
          type: boolean
    JournalPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        author:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    JournalPostInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        author:
          type: string
        published:
          type: boolean
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role or permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create an order (public checkout; totals computed server-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Orders]
      summary: List orders (section 'orders')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order (section 'orders')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Orders]
      summary: Update an order (section 'orders')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Orders]
      summary: Delete an order (section 'orders')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...
  /api/workshops/all:
    get:
      tags: [Workshops]
      summary: List all workshops incl. unpublished (section 'workshops')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All workshops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workshop'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/shop/all:
    get:
      tags: [Shop]
      summary: List all products incl. unpublished (section 'shop')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/gallery:
    get:
      tags: [Gallery]
      summary: List published gallery items (public)
      responses:
        '200':
          description: Published items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GalleryItem'
    post:
      tags: [Gallery]
      summary: Create a gallery item (section 'gallery')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryItemInput'
      responses:
        '201':
          description: Created item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/gallery/all:
    get:
      tags: [Gallery]
      summary: List all gallery items incl. unpublished (section 'gallery')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GalleryItem'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/gallery/{id}:
    patch:
      tags: [Gallery]
      summary: Update a gallery item (section 'gallery')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryItemInput'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryItem'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Gallery]
      summary: Delete a gallery item (section 'gallery')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/journal:
    get:
      tags: [Journal]
      summary: List published journal posts (public)
      responses:
        '200':
          description: Published posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JournalPost'
    post:
      tags: [Journal]
      summary: Create a journal post (section 'journal')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalPostInput'
      responses:
        '201':
          description: Created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/journal/all:
    get:
      tags: [Journal]
      summary: List all journal posts incl. unpublished (section 'journal')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JournalPost'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/journal/{id}:
    patch:
      tags: [Journal]
      summary: Update a journal post (section 'journal')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalPostInput'
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalPost'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Journal]
      summary: Delete a journal post (section 'journal')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...